Made changes on authorization

+++ controller/bootcamps.js
@@ -73,6 +73,15 @@ exports.getBootCampByID = asyncHandler(async (req, res, next) => {
 //@route POST/api/v1/bootcamps
 //access PRIVATE
 exports.postBootCamp = asyncHandler(async (req, res, next) => {
+  req.body.user = req.user.id
+
+  //Check if published bootcamps
+  const publishedBootcamps = await Bootcamp.findOne({ user: req.user.id })
+
+  if (publishedBootcamps && req.user.role === 'publisher') {
+    return next(new ErrorResponse(`The user with Id ${req.user.id} already published a bootcamp`, 400))
+  }
+
   const bootcamp = await Bootcamp.create(req.body)
   res.status(201).json({ success: true, data: bootcamp })
 })
@@ -81,10 +90,17 @@ exports.postBootCamp = asyncHandler(async (req, res, next) => {
 //@route PUT/api/v1/bootcamp/:id
 //@access PRIVATE
 exports.putBootCamp = asyncHandler(async (req, res, next) => {
-  const bootcamp = await Bootcamp.findByIdAndUpdate(req.params.id, req.body, { new: true, runValidators: true })
+  let bootcamp = await Bootcamp.findById(req.params.id)
   if (!bootcamp) {
     return next(new ErrorResponse(`Bootcamp not found with id of ${req.params.id}`, 404))
   }
+
+  //Make sure user is bootcamp owner
+  if (bootcamp.user.toString() !== req.user.id && req.user.role !== 'admin') {
+    return next(new ErrorResponse(`User ${req.user.id} is not authorized to update this bootcamp`, 401))
+  }
+
+  bootcamp = await Bootcamp.findByIdAndUpdate(req.params.id, req.body, { new: true, runValidators: true })
   res.status(200).json({ success: true, data: bootcamp })
 })
 
@@ -93,10 +109,18 @@ exports.putBootCamp = asyncHandler(async (req, res, next) => {
 //@route DELETE /api/v1/bootcamps/:id
 //@access PRIVATE 
 exports.deleteBootCamp = asyncHandler(async (req, res, next) => {
-  const bootcamp = await Bootcamp.findByIdAndDelete(req.params.id)
+  let bootcamp = await Bootcamp.findById(req.params.id)
   if (!bootcamp) {
-    return next(new ErrorResponse(`Bootcamp not found with id of ${req.params.id}`, 404))
+    return next(new ErrorResponse(`Bootcamp not found with id of ${req.params.id}`, 401))
   }
+
+  //Make sure the user is bootcamp owner
+  if (bootcamp.user.toString() !== req.user.id && req.user.role !== 'admin') {
+    return next(new ErrorResponse(`User ${req.user.id} is not authorized to delete this bootcamp`))
+  }
+
+  bootcamp = await Bootcamp.findByIdAndDelete(req.params.id)
+
   res.status(200).json({ success: true, data: 'Data deleted' })
 })
